requirements.txt
django==4.2.11
psycopg2-binary==2.9.9
ipython==8.15.0
pillow==10.2.0
djangorestframework==3.15.0
django-filter==24.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
